refactor(frontend): Applied changes according last commit
