+++ src/frontend/src/pages/login.astro
@@ -1,7 +1,7 @@
 ---
 import AuthenticationLayout from "../layouts/AuthenticationLayout.astro"
 import ErrorBanner from "../components/ErrorBanner.astro"
-import {login} from "../utils/users"
+import {login} from "../utils/server/users"
 
 let error = "";
 const credentials = {
